refactor: remove LottoNumbers and make Lotto replace LottoNumbers
